refactoring: redis cache mock in repository test
