+++ repository/comment_test.go
@@ -1,20 +1,36 @@
 package repository
 
 import (
+	rcache "github.com/go-redis/cache/v8"
+	"github.com/golang/mock/gomock"
 	"github.com/khu-dev/khumu-comment/data"
 	"github.com/khu-dev/khumu-comment/ent/enttest"
 	"github.com/khu-dev/khumu-comment/errorz"
+	"github.com/khu-dev/khumu-comment/repository/cache"
 	"github.com/khu-dev/khumu-comment/test"
 	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
+var (
+	ctrl             *gomock.Controller
+	commentCacheRepo *cache.MockCommentCacheRepository
+)
+
+func Before(t *testing.T) {
+	ctrl = gomock.NewController(t)
+	commentCacheRepo = cache.NewMockCommentCacheRepository(ctrl)
+	commentCacheRepo.EXPECT().FindAllParentCommentsByArticleID(gomock.Any()).Return(nil, rcache.ErrCacheMiss).AnyTimes()
+	commentCacheRepo.EXPECT().SetCommentsByArticleID(gomock.Any(), gomock.Any()).AnyTimes()
+}
+
 func Test_commentRepository_Create(t *testing.T) {
 	t.Run("성공", func(t *testing.T) {
+		Before(t)
 		db := enttest.Open(t, "sqlite3", "file:ent?mode=memory&cache=shared&_fk=1")
 		defer db.Close()
-		repo := NewCommentRepository(db)
+		repo := NewCommentRepository(db, commentCacheRepo)
 		test.SetUpUsers(db)
 		test.SetUpArticles(db)
 
@@ -47,9 +63,10 @@ func Test_commentRepository_Create(t *testing.T) {
 
 func TestCommentRepository_FindAllParentsByAuthorID(t *testing.T) {
 	t.Run("성공", func(t *testing.T) {
+		Before(t)
 		db := enttest.Open(t, "sqlite3", "file:ent?mode=memory&cache=shared&_fk=1")
 		defer db.Close()
-		repo := NewCommentRepository(db)
+		repo := NewCommentRepository(db, commentCacheRepo)
 		test.SetUpUsers(db)
 		test.SetUpArticles(db)
 
@@ -82,9 +99,10 @@ func TestCommentRepository_FindAllParentsByAuthorID(t *testing.T) {
 
 func TestCommentRepository_Get(t *testing.T) {
 	t.Run("ErrResourceNotFound 에러 랩핑", func(t *testing.T) {
+		Before(t)
 		db := enttest.Open(t, "sqlite3", "file:ent?mode=memory&cache=shared&_fk=1")
 		defer db.Close()
-		repo := NewCommentRepository(db)
+		repo := NewCommentRepository(db, commentCacheRepo)
 		test.SetUpUsers(db)
 		test.SetUpArticles(db)
 		// error type warpping 테스트
