+++ script/mockgen.sh
@@ -10,7 +10,7 @@
 mockgen -package usecase -destination usecase/mock.go -source usecase/comment.go
 mockgen -package repository -destination repository/comment_mock.go -source repository/comment.go
 mockgen -package repository -destination repository/likecoment_mock.go -source repository/likecomment.go
-mockgen -package repository -destination repository/cache_mock.go -source repository/cache.go
+mockgen -package repository -destination repository/cache/mock.go -source repository/cache/cache.go
 mockgen -package external -destination external/awssns_mock.go -source external/awssns.go
 mockgen -package external -destination external/redis_mock.go -source external/redis.go
 mockgen -package khumu -destination external/khumu/mock.go -source external/khumu/api.go
